feat: 모바일 헤더에 검색 input UI 적용 - SearchInput 컴포넌트 생성(#58)
