+++ src/components/layout/common/index.ts
@@ -1,2 +1,3 @@
 export { Logo } from './Logo';
 export { NavMenu } from './NavMenu';
+export { default as SearchInput } from './SearchInput';
